requirements.txt
flask
flask_RESTful
flask-cors
mysql-connector-python
validators
